Add confirmation before deletion

+++ products/views.py
@@ -141,6 +141,11 @@ def delete_product(request, product_id):
         )
         return redirect(reverse("home"))
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, "Product successfully deleted!")
-    return redirect(reverse("products"))
+    if request.method == "POST":
+        product.delete()
+        messages.success(request, "Product successfully deleted!")
+        return redirect(reverse("products"))
+    context = {
+        "product": product,
+    }
+    return render(request, "products/delete_product.html", context)
